pyproject.toml
[project]
name = "heart-disease-prediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "joblib>=1.5.3",
    "jupyter>=1.1.1",
    "kagglehub>=0.3.13",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "streamlit>=1.52.2",
]
